+++ magpie/api/schemas.py
@@ -198,6 +198,9 @@ GroupResourcePermissionsAPI = Service(
 GroupResourcePermissionAPI = Service(
     path="/groups/{group_name}/resources/{resource_id}/permissions/{permission_name}",
     name="GroupResourcePermission")
+PermissionsAPI = Service(
+    path="/permissions",
+    name="Permissions")
 RegisterGroupsAPI = Service(
     path="/register/groups",
     name="RegisterGroups")
@@ -855,6 +858,73 @@ class PermissionNameListSchema(colander.SequenceSchema):
         description="Permissions applicable to the service/resource",
         example=Permission.READ.value
     )
+
+
+class PermissionPatchObjectSchema(colander.MappingSchema):
+    resource_name = colander.SchemaNode(
+        colander.String(),
+        description="Name of the resource associated with the permission. This resource will be created if missing."
+    )
+    resource_type = colander.SchemaNode(
+        colander.String(),
+        description="Type of the resource. The first resource must be of the `service` type, and children "
+                    "resources must have a relevant resource type that is not of the `service` type.",
+        example="service"
+    )
+
+    user = colander.SchemaNode(
+        colander.String(),
+        description="Registered local user associated with the permission.",
+        example="toto",
+        missing=colander.drop
+    )
+    group = colander.SchemaNode(
+        colander.String(),
+        description="Registered user group associated with the permission.",
+        example="users",
+        missing=colander.drop
+    )
+
+    # FIXME: support oneOf(string, object), permission can actually be either a string or a PermissionObjectSchema(dict)
+    # Currently not possible to have multiple type options with colander.
+    permission = PermissionObjectSchema(
+        missing=colander.drop
+    )
+
+    action = colander.SchemaNode(
+        colander.String(),
+        description="Action to apply on the permission.",
+        example="create",
+        default="create",
+        validator=colander.OneOf(["create", "remove"])
+    )
+
+
+class PermissionPatchListSchema(colander.SequenceSchema):
+    permission = PermissionPatchObjectSchema()
+
+
+class Permissions_PATCH_RequestBodySchema(colander.MappingSchema):
+    permissions = PermissionPatchListSchema()
+
+
+class Permissions_PATCH_RequestSchema(BaseRequestSchemaAPI):
+    body = Permissions_PATCH_RequestBodySchema()
+
+
+class Permissions_PATCH_OkResponseSchema(BaseResponseSchemaAPI):
+    description = "Update permissions successful."
+    body = BaseResponseBodySchema(code=HTTPOk.code, description=description)
+
+
+class Permissions_PATCH_BadRequestResponseSchema(BaseResponseSchemaAPI):
+    description = "Missing or invalid parameters for permissions update."
+    body = ErrorResponseBodySchema(code=HTTPBadRequest.code, description=description)
+
+
+class Permissions_PATCH_ForbiddenResponseSchema(BaseResponseSchemaAPI):
+    description = "Failed to update requested permissions."
+    body = ErrorResponseBodySchema(code=HTTPForbidden.code, description=description)
 
 
 class BaseUserInfoSchema(colander.MappingSchema):
@@ -3589,6 +3659,14 @@ ServiceResource_DELETE_responses = {
     "404": Resource_MatchDictCheck_NotFoundResponseSchema(),
     "406": NotAcceptableResponseSchema(),
     "422": UnprocessableEntityResponseSchema(),
+    "500": InternalServerErrorResponseSchema(),
+}
+Permissions_PATCH_responses = {
+    "200": Permissions_PATCH_OkResponseSchema(),
+    "400": Permissions_PATCH_BadRequestResponseSchema(),  # FIXME: https://github.com/Ouranosinc/Magpie/issues/359
+    "401": UnauthorizedResponseSchema(),
+    "403": Permissions_PATCH_ForbiddenResponseSchema(),
+    "406": NotAcceptableResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
 Users_GET_responses = {